+++ hiveeyes/datalogger.py
@@ -66,6 +66,7 @@ class HiveeyesDatalogger(TerkinDatalogger):
         settings = self.settings.get('sensors.registry.hx711')
 
         hx711_sensor = HX711Sensor()
+        hx711_sensor.set_address(settings.get('address', 0x00))
         hx711_sensor.register_pin('dout', settings['pin_dout'])
         hx711_sensor.register_pin('pdsck', settings['pin_pdsck'])
         hx711_sensor.register_parameter('scale', settings['scale'])